moved app shell factory to app.
added output data obejct to preferenfces,a nd updated relevant files to work with the changes.
