+++ src/main/java/interface_adapter/preferences/GetPreferencesPresenter.java
@@ -2,6 +2,7 @@ package interface_adapter.preferences;
 
 import entity.Preferences;
 import use_case.preferences.get_preferences.GetPreferencesOutputBoundary;
+import use_case.preferences.get_preferences.GetPreferencesOutputData;
 
 public class GetPreferencesPresenter implements GetPreferencesOutputBoundary {
     private final PreferencesViewModel viewModel;
@@ -11,7 +12,7 @@ public class GetPreferencesPresenter implements GetPreferencesOutputBoundary {
     }
 
     @Override
-    public void presentPreferences(Preferences preferences) {
-        viewModel.setPreferences(preferences);
+    public void presentPreferences(GetPreferencesOutputData preferences) {
+        viewModel.setPreferences(preferences.getPreferences());
     }
 }
