+++ src/main/java/use_case/preferences/get_preferences/GetPreferencesOutputBoundary.java
@@ -3,5 +3,5 @@ package use_case.preferences.get_preferences;
 import entity.Preferences;
 
 public interface GetPreferencesOutputBoundary {
-    void presentPreferences(Preferences preferences);
+    void presentPreferences(GetPreferencesOutputData outputData );
 }
